requirements.txt
aenum==3.1.11
asn1tools==0.163.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bitstruct==8.15.1
boltons==21.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorlog==6.6.0
dacite==1.6.0
decorator==5.1.1
diskcache==5.4.0
executing==0.8.3
future==0.18.2
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.22.4
parrot-olympe==7.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.14.1
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
stack-data==0.2.0
traitlets==5.2.2.post1
tzlocal==3.0
urllib3==1.26.9
wcwidth==0.2.5
wsproto==1.1.0
yapf==0.32.0
